// src/app/models/registration.model.ts
export class Registration {
    CID:number;
    /* title: string; */
	FirstName: string;
	MiddleName: string;
	LastName: string;
	EmailID: string;
	MobileNumber: string;
    FatherName: string;
    AadharNumber:number;
	DOB: Date;
	Occupationtype: string;
	SourceofIncome: string;
	GrossannualIncome: number;
	DebitCardOpted: boolean; 
	NetBankingOpted: boolean; 
    RAddressLine1: string;
    RAddressLine2: string;
    RLandmark: string;
    RState: string;
    RCity: string;
    RPincode: string;
    PAddressLine1: string;
    PAddressLine2: string;
    PLandmark: string;
    PState: string;
    PCity: string;
    PPincode: string;
}